clazy++: Qt6::qhash signature has changed from uint to size_t

+++ core/libs/widgets/itemview/ditemdelegate.cpp
@@ -199,7 +199,7 @@ QString DItemDelegate::squeezedTextCached(QPainter* const p, int width, const QS
 
     // We do not need to include the font into cache key, the cache is cleared on font change
 
-    QString cacheKey                      = QString::number(width) + QString::number((uint)qHash(text));
+    QString cacheKey                      = QString::number(width) + QString::number(static_cast<qulonglong>(qHash(text)));
     QString* const cachedString           = cache->object(cacheKey);
 
     if (cachedString)
